Refactor compiler into helper classes
