+++ src/magma/ast/Invocation.java
@@ -1,7 +1,7 @@
 package magma.ast;
 
 import magma.util.*;
-import magma.Main;
+import magma.Generator;
 import magma.compile.*;
 /**
  * Invocation of a {@link Caller} with a list of argument values.
@@ -9,6 +9,6 @@ import magma.compile.*;
 public record Invocation(Caller caller, List<Value> arguments) implements Value {
     @Override
     public String generate() {
-        return caller.generate() + "(" + Main.generateNodes(arguments) + ")";
+        return caller.generate() + "(" + Generator.generateNodes(arguments) + ")";
     }
 }
