+++ src/magma/ast/TemplateType.java
@@ -1,12 +1,12 @@
 package magma.ast;
 
 import magma.util.*;
-import magma.Main;
+import magma.Generator;
 import magma.compile.*;
 public record TemplateType(String base, List<Type> arguments) implements Type {
     @Override
     public String generate() {
-        final var outputArguments = Main.generateNodes(arguments);
+        final var outputArguments = Generator.generateNodes(arguments);
         return base + "<" + outputArguments + ">";
     }
 }
