+++ test/magma/RecordTypeParamTest.java
@@ -12,7 +12,7 @@ public class RecordTypeParamTest {
         TypeParam typeParam = new TypeParam("T");
         TypeParamSet typeParams = new TypeParamSet(Lists.of(typeParam));
         CompileState state = new CompileState().enter(new StructureFrame(typeParams));
-        Parameter param = Main.parseParameter("T value", state);
+        Parameter param = Parser.parseParameter("T value", state);
         assertTrue(param instanceof Definition);
         Definition def = (Definition) param;
         assertTrue(def.type instanceof TypeParam, "expected type param to resolve in parameter");
